Refactor: 권유 모달로 리펙토링
