+++ src/storybook/components/RecommendModal.stories.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
-import { Typography } from '@mui/material';
-import AlertPWA from '@/components/AlertPWA';
+import RecommendModal from '@/components/RecommendModal';
+import PWAModalContent from '@/pages/main/PWAModalContent';
 
 const meta = {
-    title: 'components/AlertPWA',
-    component: AlertPWA,
+    title: 'components/RecommendModal',
+    component: RecommendModal,
     tags: ['autodocs'],
     argTypes: {
         open: { control: 'boolean', description: '모달 open 여부' },
@@ -20,13 +20,7 @@ const Default = {
         onClose: () => {}, // onClose 함수는 비어있는 함수로 설정
         buttonText: '확인', // 버튼 텍스트는 '확인'으로 설정
         onButtonClick: () => {}, // onButtonClick 함수는 비어있는 함수로 설정
-        children: (
-            <Typography align='center' variant='title2'>
-                보관 후에는 수정이 불가해요.
-                <br />
-                이대로 전달할까요?
-            </Typography>
-        ),
+        children: <PWAModalContent />,
     },
 };
